requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via gcsfs
aiosignal==1.3.1
    # via aiohttp
altair==5.3.0
    # via streamlit
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.3.3
    # via
    #   google-auth
    #   streamlit
certifi==2024.6.2
    # via
    #   pinecone-client
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
decorator==5.1.1
    # via gcsfs
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.0
    # via
    #   gcsfs
    #   huggingface-hub
    #   torch
gcsfs==2024.6.0
    # via -r requirements.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-api-core==2.19.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.30.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-oauthlib==1.2.0
    # via gcsfs
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.17.0
    # via gcsfs
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.1
    # via google-cloud-storage
googleapis-common-protos==1.63.1
    # via google-api-core
huggingface-hub==0.23.4
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.4.2
    # via scikit-learn
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
networkx==3.3
    # via torch
numpy==2.0.0
    # via
    #   -r requirements.in
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   streamlit
    #   transformers
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.1
    # via
    #   altair
    #   huggingface-hub
    #   streamlit
    #   transformers
pandas==2.2.2
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pillow==10.3.0
    # via
    #   sentence-transformers
    #   streamlit
pinecone-client==4.1.1
    # via -r requirements.in
pinecone-plugin-interface==0.0.7
    # via pinecone-client
proto-plus==1.24.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
pyarrow==16.1.0
    # via streamlit
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via transformers
requests==2.32.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-storage
    #   huggingface-hub
    #   requests-oauthlib
    #   streamlit
    #   transformers
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.7.1
    # via streamlit
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
safetensors==0.4.3
    # via transformers
scikit-learn==1.5.0
    # via sentence-transformers
scipy==1.13.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentencepiece
sentence-transformers==3.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.36.0
    # via -r requirements.in
sympy==1.12.1
    # via torch
tenacity==8.4.1
    # via streamlit
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via altair
torch==2.3.1
    # via sentence-transformers
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   pinecone-client
    #   sentence-transformers
    #   transformers
transformers==4.41.2
    # via sentence-transformers
triton==2.3.1
    # via torch
typing-extensions==4.12.2
    # via
    #   altair
    #   huggingface-hub
    #   pinecone-client
    #   streamlit
    #   torch
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   pinecone-client
    #   requests
watchdog==4.0.1
    # via streamlit
yarl==1.9.4
    # via aiohttp
